Profile: new option for flat recursion
Under ragged recursion, the tree profile can be difficult to interpret
since it becomes unclear whether time is being spent in the recursion
or doing work for the function itself. To separate those effects,
we can subtract off the cost of the recursion from the accumulated sum
by rearranging the tree to only include the unique (non-recursive)
subgraphs. In many cases, this should be similar in effect to re-writting
all recursive functions as iterators.

This is enabled with the new option :recur \in (:off, :flat, :flatc)
for the format=:tree view. Implementing it for the :flat view is left
as an exercise for the reader (although it is believed to be much easier).